uses the directory's atime equal to 0 to identify whether it is an existing S3 object

+++ pkg/gateway/gateway.go
@@ -49,11 +49,10 @@ var mctx meta.Context
 var logger = utils.GetLogger("juicefs")
 
 type Config struct {
-	MultiBucket     bool
-	KeepEtag        bool
-	Mode            uint16
-	DirMode         uint16
-	ReturnDirInList bool
+	MultiBucket bool
+	KeepEtag    bool
+	Mode        uint16
+	DirMode     uint16
 }
 
 func NewJFSGateway(jfs *fs.FileSystem, conf *vfs.Config, gConf *Config) (minio.ObjectLayer, error) {
@@ -282,9 +281,10 @@ func (n *jfsObjects) isLeaf(bucket, leafPath string) bool {
 	return !strings.HasSuffix(leafPath, "/")
 }
 
-func (n *jfsObjects) listDirFactory(delimiter string) minio.ListDirFunc {
+func (n *jfsObjects) listDirFactory() minio.ListDirFunc {
 	return func(bucket, prefixDir, prefixEntry string) (emptyDir bool, entries []string, delayIsLeaf bool) {
-		if n.gConf.ReturnDirInList && delimiter != sep && prefixEntry == "" {
+		// if the prefixEntry is empty, and the atime of prefixDir is 0, return it in the result
+		if prefixEntry == "" {
 			entries = append(entries, "")
 		}
 		f, eno := n.fs.Open(mctx, n.path(bucket, prefixDir), 0)
@@ -296,9 +296,6 @@ func (n *jfsObjects) listDirFactory(delimiter string) minio.ListDirFunc {
 		if eno != 0 {
 			return
 		}
-		if len(fis) == 0 {
-			return true, nil, false
-		}
 		root := n.path(bucket, prefixDir) == "/"
 		for _, fi := range fis {
 			if root && len(fi.Name()) == len(metaBucket) && fi.Name() == metaBucket {
@@ -309,6 +306,9 @@ func (n *jfsObjects) listDirFactory(delimiter string) minio.ListDirFunc {
 			} else {
 				entries = append(entries, fi.Name())
 			}
+		}
+		if len(entries) == 0 {
+			return true, nil, false
 		}
 		entries, delayIsLeaf = minio.FilterListEntries(bucket, prefixDir, entries, prefixEntry, n.isLeaf)
 		return false, entries, delayIsLeaf
@@ -343,7 +343,7 @@ func (n *jfsObjects) ListObjects(ctx context.Context, bucket, prefix, marker, de
 				ModTime: fi.ModTime(),
 				Size:    size,
 				IsDir:   fi.IsDir(),
-				AccTime: fi.ModTime(),
+				AccTime: fi.ACCTime(),
 			}
 		}
 		return obj, jfsToObjectErr(ctx, eno, bucket, object)
@@ -352,7 +352,7 @@ func (n *jfsObjects) ListObjects(ctx context.Context, bucket, prefix, marker, de
 	if maxKeys == 0 {
 		maxKeys = -1 // list as many objects as possible
 	}
-	return minio.ListObjects(ctx, n, bucket, prefix, marker, delimiter, maxKeys, n.listPool, n.listDirFactory(delimiter), n.isLeaf, n.isLeafDir, getObjectInfo, getObjectInfo)
+	return minio.ListObjects(ctx, n, bucket, prefix, marker, delimiter, maxKeys, n.listPool, n.listDirFactory(), n.isLeaf, n.isLeafDir, getObjectInfo, getObjectInfo)
 }
 
 // ListObjectsV2 lists all blobs in JFS bucket filtered by prefix
@@ -379,6 +379,34 @@ func (n *jfsObjects) ListObjectsV2(ctx context.Context, bucket, prefix, continua
 	return loi, err
 }
 
+func (n *jfsObjects) setZeroAtime(p string) bool {
+	if f, eno := n.fs.Open(mctx, p, 0); eno == 0 {
+		defer f.Close(mctx)
+		if fis, eno := f.Readdir(mctx, 0); eno == 0 {
+			if len(fis) != 0 {
+				if fi, errno := n.fs.Stat(mctx, p); errno == 0 {
+					var attr = &meta.Attr{}
+					now := time.Now()
+					attr.Atime = now.Unix()
+					attr.Atimensec = uint32(now.Nanosecond())
+					if errno := n.fs.Meta().SetAttr(mctx, fi.Inode(), meta.SetAttrAtime, 0, attr); errno == 0 {
+						return true
+					} else {
+						logger.Warnf("set dir atime failed %s: %s", p, errno)
+					}
+				} else {
+					logger.Warnf("stat dir failed %s: %s", p, errno)
+				}
+			}
+		} else {
+			logger.Warnf("readdir: %s", eno)
+		}
+	} else {
+		logger.Warnf("open: %s", eno)
+	}
+	return false
+}
+
 func (n *jfsObjects) DeleteObject(ctx context.Context, bucket, object string, options minio.ObjectOptions) (info minio.ObjectInfo, err error) {
 	if err = n.checkBucket(ctx, bucket); err != nil {
 		return
@@ -387,16 +415,26 @@ func (n *jfsObjects) DeleteObject(ctx context.Context, bucket, object string, op
 	info.Name = object
 	p := n.path(bucket, object)
 	root := n.path(bucket)
-	for p != root {
-		if eno := n.fs.Delete(mctx, p); eno != 0 {
-			if fs.IsNotEmpty(eno) {
-				err = nil
-			} else {
-				err = eno
+
+	// if the delete object is a directory, and it has children entry, set its atime to now to represent that it has been removed
+	needDelete := true
+	if strings.HasSuffix(object, sep) {
+		needDelete = !n.setZeroAtime(n.path(bucket, object))
+	}
+	// p should not end with /  because if p=/d1/d2/  next step p is /d1/d2 instead of /d1 as expected
+	p = path.Clean(p)
+	if needDelete {
+		for p != root {
+			if eno := n.fs.Delete(mctx, p); eno != 0 {
+				if fs.IsNotEmpty(eno) {
+					err = nil
+				} else {
+					err = eno
+				}
+				break
 			}
-			break
+			p = path.Dir(p)
 		}
-		p = path.Dir(p)
 	}
 	return info, jfsToObjectErr(ctx, err, bucket, object)
 }
@@ -504,7 +542,7 @@ func (n *jfsObjects) CopyObject(ctx context.Context, srcBucket, srcObject, dstBu
 		ModTime: fi.ModTime(),
 		Size:    fi.Size(),
 		IsDir:   fi.IsDir(),
-		AccTime: fi.ModTime(),
+		AccTime: fi.ACCTime(),
 	}, nil
 }
 
@@ -575,7 +613,7 @@ func (n *jfsObjects) GetObjectInfo(ctx context.Context, bucket, object string, o
 		ModTime: fi.ModTime(),
 		Size:    size,
 		IsDir:   fi.IsDir(),
-		AccTime: fi.ModTime(),
+		AccTime: fi.ACCTime(),
 		ETag:    string(etag),
 	}, nil
 }
@@ -663,6 +701,20 @@ func (n *jfsObjects) PutObject(ctx context.Context, bucket string, object string
 			err = jfsToObjectErr(ctx, err, bucket, object)
 			return
 		}
+
+		// if the put object is a directory, set its atime to 0
+		if fi, errno := n.fs.Stat(mctx, p); errno == 0 {
+			var attr = &meta.Attr{}
+			zeroTime := time.Time{}
+			attr.Atime = zeroTime.Unix()
+			attr.Atimensec = uint32(zeroTime.Nanosecond())
+			if errno := n.fs.Meta().SetAttr(mctx, fi.Inode(), meta.SetAttrAtime, 0, attr); errno != 0 {
+				logger.Warnf("set dir atime failed %s: %s", p, errno)
+			}
+		} else {
+			logger.Warnf("stat dir failed %s: %s", p, errno)
+		}
+
 		if r.Size() > 0 {
 			err = minio.ObjectExistsAsDirectory{
 				Bucket: bucket,
@@ -692,7 +744,7 @@ func (n *jfsObjects) PutObject(ctx context.Context, bucket string, object string
 		ModTime: fi.ModTime(),
 		Size:    fi.Size(),
 		IsDir:   fi.IsDir(),
-		AccTime: fi.ModTime(),
+		AccTime: fi.ACCTime(),
 	}, nil
 }
 
@@ -921,7 +973,7 @@ func (n *jfsObjects) CompleteMultipartUpload(ctx context.Context, bucket, object
 		ModTime: fi.ModTime(),
 		Size:    fi.Size(),
 		IsDir:   fi.IsDir(),
-		AccTime: fi.ModTime(),
+		AccTime: fi.ACCTime(),
 	}, nil
 }
